+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -1,52 +1,12 @@
 package com.techcourse;
 
-import com.interface21.web.WebApplicationInitializer;
-import com.interface21.webmvc.servlet.mvc.adapter.HandlerExecutionHandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.mapping.AnnotationHandlerMapping;
-import jakarta.servlet.ServletContext;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.interface21.web.AbstractDispatcherServletInitializer;
 
-/**
- * Base class for {@link WebApplicationInitializer}
- * implementations that register a {@link DispatcherServlet} in the servlet context.
- */
-public class DispatcherServletInitializer implements WebApplicationInitializer {
+public class DispatcherServletInitializer extends AbstractDispatcherServletInitializer {
 
-    private static final Logger log = LoggerFactory.getLogger(DispatcherServletInitializer.class);
-
-    private static final String DEFAULT_SERVLET_NAME = "dispatcher";
 
     @Override
-    public void onStartup(final ServletContext servletContext) {
-
-        final var dispatcherServlet = new DispatcherServlet(createHandlerMappingRegistry(), createHandlerAdaptorRegistry());
-
-        final var registration = servletContext.addServlet(DEFAULT_SERVLET_NAME, dispatcherServlet);
-        if (registration == null) {
-            throw new IllegalStateException("Failed to register servlet with name '" + DEFAULT_SERVLET_NAME + "'. " +
-                    "Check if there is another servlet registered under the same name.");
-        }
-
-        registration.setLoadOnStartup(1);
-        registration.addMapping("/");
-
-        log.info("Start AppWebApplication Initializer");
-    }
-
-    private HandlerMappingRegistry createHandlerMappingRegistry() {
-        HandlerMappingRegistry registry = new HandlerMappingRegistry();
-
-        AnnotationHandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping("com.techcourse.controller");
-        annotationHandlerMapping.initialize();
-
-        registry.addHandlerMapping(annotationHandlerMapping);
-        return registry;
-    }
-
-    private HandlerAdaptorRegistry createHandlerAdaptorRegistry() {
-        HandlerAdaptorRegistry registry = new HandlerAdaptorRegistry();
-        registry.addHandlerAdapter(new HandlerExecutionHandlerAdapter());
-        return registry;
+    protected String getBasePackage() {
+        return "com.techcourse.controller";
     }
 }
