+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/DispatcherServlet.java
@@ -1,8 +1,10 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.View;
 import com.interface21.webmvc.servlet.mvc.adapter.HandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.adapter.HandlerExecutionHandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.mapping.AnnotationHandlerMapping;
 import com.interface21.webmvc.servlet.mvc.mapping.HandlerMapping;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
@@ -17,17 +19,18 @@ public class DispatcherServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
-    private final HandlerMappingRegistry handlerMappingRegistry;
-    private final HandlerAdaptorRegistry handlerAdaptorRegistry;
+    private final HandlerMappingRegistry handlerMappingRegistry = new HandlerMappingRegistry();
+    private final HandlerAdaptorRegistry handlerAdaptorRegistry = new HandlerAdaptorRegistry();
+    private final Object[] basePackages;
 
-    public DispatcherServlet(HandlerMappingRegistry handlerMappingRegistry,
-                             HandlerAdaptorRegistry handlerAdaptorRegistry) {
-        this.handlerMappingRegistry = handlerMappingRegistry;
-        this.handlerAdaptorRegistry = handlerAdaptorRegistry;
+    public DispatcherServlet(final Object... basePackages) {
+        this.basePackages = basePackages;
     }
 
     @Override
     public void init() {
+        initHandlerMappingRegistry();
+        initHandlerAdaptorRegistry();
     }
 
     @Override
@@ -54,5 +57,16 @@ public class DispatcherServlet extends HttpServlet {
             log.error("Exception : {}", e.getMessage(), e);
             throw new ServletException(e.getMessage());
         }
+    }
+
+    private void initHandlerMappingRegistry() {
+        AnnotationHandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping(basePackages);
+        annotationHandlerMapping.initialize();
+
+        handlerMappingRegistry.addHandlerMapping(annotationHandlerMapping);
+    }
+
+    private void initHandlerAdaptorRegistry() {
+        handlerAdaptorRegistry.addHandlerAdapter(new HandlerExecutionHandlerAdapter());
     }
 }
